Organize classes to be in their own files.
