+++ src/Imageflow/Internal.Helpers/OutputDestinationWrapper.cs
